+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * started returning resource id in Mute / hold / transfers publications
 * added appending plus sign to sms phone numbers
 * replace the fb id with the fb name in the "from" field of facebook messaging interactions
+* added resource check capacity functionality
+* added realtime stats polling and available stats
 * added permissions checking at tenant set
 * added wrapup functionality
 * added extension setting functionality